requirements.txt
numpy>=1.24.0
torch>=2.0.0
tensorboard
termcolor
highway_env
matplotlib
PyYAML
imageio
imageio-ffmpeg
attrdict